Using NSValueTransformer to provide NSAttributedStrings for controls isn't as clever as it sounds
CCMProjectDefaultValueTransformer provides an NSAttributedString for displaying cells/rows in the "projects" table view within preferences. However, highlighting the (view based) cells has no effect on the displayed text, as the value transformer doesn't have any relation to the context the text will be displayed in. Highlighting cells thus doesn't lead to changing the text attributes in a sensible way.

This hack provides a way for the value transformer to create a proper attributed string.